pyproject.toml
[project]
name = "wallet-attached-storage-server"
version = "0.1.0"
description = "Server implementation for the Wallet Attached Storage specification"
license = { file = "LICENSE" }
readme = "README.md"
authors = [
    { name = "Grey Flannel Consulting", email = "contact@greyflannelconsulting.com" }
]
requires-python = ">=3.11"
dependencies = [
    "base58>=2.1",
    "cryptography>=46.0.5",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[project.optional-dependencies]
postgresql = ["psycopg[binary]>=3.1", "psycopg-pool>=3.1"]
s3 = ["boto3>=1.26"]
onedrive = ["msal>=1.20", "httpx>=0.27"]
dropbox = ["dropbox>=12.0"]
gdrive = ["google-api-python-client>=2.100", "google-auth>=2.20"]
all-backends = ["wallet-attached-storage-server[postgresql,s3,onedrive,dropbox,gdrive]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "flake8>=7.3.0",
    "httpx>=0.27",
    "keyring>=25.6.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
    "safety>=3.2.9",
    "moto[s3]>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/was_server"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
markers = [
    "live: tests requiring a WAS server on localhost:8080",
    "postgresql: tests requiring a PostgreSQL database",
    "s3: tests for S3 backend (uses moto mock)",
    "onedrive: tests for OneDrive backend",
    "dropbox: tests for Dropbox backend",
    "gdrive: tests for Google Drive backend",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "N", "S"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]
